+++ R/formatage.annee.neutre.R
@@ -18,7 +18,7 @@ formatage.annee.neutre <- function(
   #### Calculs ####
   data <- 
     data %>% 
-    mutate(chmes_date_anneeneutre = ifelse(str_sub(chmes_date, 6, 10) >= datedebutanneeneutre, format(chmes_date, "2000-%m-%d"), format(chmes_date, "2001-%m-%d"))) %>% # Attention de bien se caler sur une année bisextile, sinon on peut générer des "2001-02-29" qui n'existent pas, et qui posent donc des problèmes de parsing avec ymd et des warnings et des NA
+    mutate(chmes_date_anneeneutre = ifelse(str_sub(chmes_date, 6, 10) >= datedebutanneeneutre, format(chmes_date, "2000-%m-%d"), format(chmes_date, "2001-%m-%d"))) %>% # Attention de bien se caler sur une année bissextile, sinon on peut générer des "2001-02-29" qui n'existent pas, et qui posent donc des problèmes de parsing avec ymd et des warnings et des NA
     filter(chmes_date_anneeneutre != "2001-02-29") %>% # On ne peut pas faire autrement, car si on met 2000 et 2004 par exemple, alors on génère des vides dans les figures utilisées ensuite : les deux années doivent se suivre, donc il y aura forcément des valeurs impossibles sur la première ou bien sur la deuxième année. On est donc obligé de filtrer d'un côté. Ces données seraient de toute manière supprimées du fait du bug de parsing qui générerait alors des NA qui ne seraient pas représentés.
     mutate(chmes_date_anneeneutre = ymd(chmes_date_anneeneutre))